+++ src/rex.sms/src/rex/sms/provider/__init__.py
@@ -4,6 +4,7 @@
 
 
 from .base import *
+from .logging import *
 from .null import *
 from .stdout import *
 from .twilioapi import *